feat(ui): astro typography tag components with token-level props
Adds individual Astro tag components (H1-H6, P, Small, Code, Blockquote,
CodeBlock, Mark, Abbr) matching the React API from #1177, minus editable
mode which requires client JS. Extracts shared tokenPropsToClasses to
typography.classes.ts for both frameworks.
