+++ packages/ui/src/components/ui/typography-blockquote.astro
@@ -1,0 +1,26 @@
+---
+/**
+ * Blockquote - Block quotation (Astro). Zero client JavaScript.
+ * For quotations or callouts with left border emphasis.
+ *
+ * @cognitive-load 1/10
+ */
+import type { HTMLAttributes } from 'astro/types';
+import classy from '../../primitives/classy';
+import { tokenPropsToClasses, typographyClasses } from './typography.classes';
+
+interface Props extends HTMLAttributes<'blockquote'> {
+  size?: string;
+  weight?: string;
+  color?: string;
+  line?: string;
+  tracking?: string;
+  family?: string;
+}
+
+const { size, weight, color, line, tracking, family, class: className, ...attrs } = Astro.props;
+const overrides = tokenPropsToClasses({ size, weight, color, line, tracking, family });
+const classes = classy(typographyClasses.blockquote, overrides, className);
+---
+
+<blockquote class={classes} {...attrs}><slot /></blockquote>
