+++ packages/ui/src/components/ui/typography-h1.astro
@@ -1,0 +1,29 @@
+---
+/**
+ * H1 - Page title (Astro). Zero client JavaScript.
+ * Use once per page for the main title.
+ *
+ * Token props override individual typography dimensions:
+ *   <H1 size="3xl" weight="black">Custom title</H1>
+ *
+ * @cognitive-load 1/10
+ */
+import type { HTMLAttributes } from 'astro/types';
+import classy from '../../primitives/classy';
+import { tokenPropsToClasses, typographyClasses } from './typography.classes';
+
+interface Props extends HTMLAttributes<'h1'> {
+  size?: string;
+  weight?: string;
+  color?: string;
+  line?: string;
+  tracking?: string;
+  family?: string;
+}
+
+const { size, weight, color, line, tracking, family, class: className, ...attrs } = Astro.props;
+const overrides = tokenPropsToClasses({ size, weight, color, line, tracking, family });
+const classes = classy(typographyClasses.h1, overrides, className);
+---
+
+<h1 class={classes} {...attrs}><slot /></h1>
