+++ packages/ui/src/components/ui/typography-h4.astro
@@ -1,0 +1,25 @@
+---
+/**
+ * H4 - Minor heading (Astro). Zero client JavaScript.
+ *
+ * @cognitive-load 1/10
+ */
+import type { HTMLAttributes } from 'astro/types';
+import classy from '../../primitives/classy';
+import { tokenPropsToClasses, typographyClasses } from './typography.classes';
+
+interface Props extends HTMLAttributes<'h4'> {
+  size?: string;
+  weight?: string;
+  color?: string;
+  line?: string;
+  tracking?: string;
+  family?: string;
+}
+
+const { size, weight, color, line, tracking, family, class: className, ...attrs } = Astro.props;
+const overrides = tokenPropsToClasses({ size, weight, color, line, tracking, family });
+const classes = classy(typographyClasses.h4, overrides, className);
+---
+
+<h4 class={classes} {...attrs}><slot /></h4>
