+++ packages/ui/src/components/ui/typography-h5.astro
@@ -1,0 +1,26 @@
+---
+/**
+ * H5 - Lower-level heading (Astro). Zero client JavaScript.
+ * Shares visual treatment with H4.
+ *
+ * @cognitive-load 1/10
+ */
+import type { HTMLAttributes } from 'astro/types';
+import classy from '../../primitives/classy';
+import { tokenPropsToClasses, typographyClasses } from './typography.classes';
+
+interface Props extends HTMLAttributes<'h5'> {
+  size?: string;
+  weight?: string;
+  color?: string;
+  line?: string;
+  tracking?: string;
+  family?: string;
+}
+
+const { size, weight, color, line, tracking, family, class: className, ...attrs } = Astro.props;
+const overrides = tokenPropsToClasses({ size, weight, color, line, tracking, family });
+const classes = classy(typographyClasses.h4, overrides, className);
+---
+
+<h5 class={classes} {...attrs}><slot /></h5>
