+++ packages/ui/src/components/ui/typography-mark.astro
@@ -1,0 +1,26 @@
+---
+/**
+ * Mark - Highlighted text (Astro). Zero client JavaScript.
+ * For search results, emphasized terms, or important callouts.
+ *
+ * @cognitive-load 1/10
+ */
+import type { HTMLAttributes } from 'astro/types';
+import classy from '../../primitives/classy';
+import { tokenPropsToClasses, typographyClasses } from './typography.classes';
+
+interface Props extends HTMLAttributes<'mark'> {
+  size?: string;
+  weight?: string;
+  color?: string;
+  line?: string;
+  tracking?: string;
+  family?: string;
+}
+
+const { size, weight, color, line, tracking, family, class: className, ...attrs } = Astro.props;
+const overrides = tokenPropsToClasses({ size, weight, color, line, tracking, family });
+const classes = classy(typographyClasses.mark, overrides, className);
+---
+
+<mark class={classes} {...attrs}><slot /></mark>
